+++ baybe/parameters/__init__.py
@@ -2,6 +2,11 @@
 
 from baybe.parameters.categorical import CategoricalParameter, TaskParameter
 from baybe.parameters.custom import CustomDiscreteParameter
+from baybe.parameters.enum import (
+    CategoricalEncoding,
+    CustomEncoding,
+    SubstanceEncoding,
+)
 from baybe.parameters.numerical import (
     NumericalContinuousParameter,
     NumericalDiscreteParameter,
@@ -9,10 +14,13 @@ from baybe.parameters.numerical import (
 from baybe.parameters.substance import SubstanceParameter
 
 __all__ = [
+    "CategoricalEncoding",
     "CategoricalParameter",
     "CustomDiscreteParameter",
+    "CustomEncoding",
     "NumericalContinuousParameter",
     "NumericalDiscreteParameter",
+    "SubstanceEncoding",
     "SubstanceParameter",
     "TaskParameter",
 ]